Rename the ArticleSearch to ArchiveSearch
HCRC-12. ArticleSearch renamed to archive search, since it's using the layout designed for articles archive search, but it was wanted to be more generic - not just for articles but also for pages, etc.
