+++ src/core/archiveSearchPageContent/archiveSearchPageContent.module.scss
